join codegen_error and diagnostic
