+++ src/codegen/generators/llvm.rs
@@ -1,5 +1,5 @@
 // Copyright (c) 2020 Ghaith Hachem and Mathias Rieder
-use crate::{ast::SourceRange, compile_error::CompileError};
+use crate::{ast::SourceRange, diagnostics::Diagnostic};
 use inkwell::{
     builder::Builder,
     context::Context,
@@ -87,7 +87,7 @@ impl<'a> Llvm<'a> {
         pointer_to_array_instance: PointerValue<'a>,
         accessor_sequence: &[IntValue<'a>],
         name: &str,
-    ) -> Result<PointerValue<'a>, CompileError> {
+    ) -> Result<PointerValue<'a>, Diagnostic> {
         unsafe {
             Ok(self
                 .builder
@@ -100,19 +100,19 @@ impl<'a> Llvm<'a> {
     /// - `pointer_to_struct_instance` a pointer to the struct
     /// - `member_index` the index of the member we want a pointer to
     /// - `name` the name of the temporary variable
-    /// - `offset` the location in case of a CompileError
+    /// - `offset` the location in case of a Diagnostic
     pub fn get_member_pointer_from_struct(
         &self,
         pointer_to_struct_instance: PointerValue<'a>,
         member_index: u32,
         name: &str,
         offset: &SourceRange,
-    ) -> Result<PointerValue<'a>, CompileError> {
+    ) -> Result<PointerValue<'a>, Diagnostic> {
         self.builder
             .build_struct_gep(pointer_to_struct_instance, member_index, name)
             .map_err(|_| {
-                CompileError::codegen_error(
-                    format!("Cannot generate qualified reference for {:}", name),
+                Diagnostic::codegen_error(
+                    &format!("Cannot generate qualified reference for {:}", name),
                     offset.clone(),
                 )
             })
@@ -142,7 +142,7 @@ impl<'a> Llvm<'a> {
     ///
     /// - `index` the index to obtain the datatypeinformation for BOOL
     /// - `value` the value of the constant bool value
-    pub fn create_const_bool(&self, value: bool) -> Result<BasicValueEnum<'a>, CompileError> {
+    pub fn create_const_bool(&self, value: bool) -> Result<BasicValueEnum<'a>, Diagnostic> {
         let itype = self.context.bool_type();
         let value = itype.const_int(value as u64, false);
         Ok(BasicValueEnum::IntValue(value))
@@ -158,19 +158,19 @@ impl<'a> Llvm<'a> {
         target_type: &BasicTypeEnum<'a>,
         value: &str,
         location: SourceRange,
-    ) -> Result<BasicValueEnum<'a>, CompileError> {
+    ) -> Result<BasicValueEnum<'a>, Diagnostic> {
         match target_type {
             BasicTypeEnum::IntType { 0: int_type } => int_type
                 .const_int_from_string(value, StringRadix::Decimal)
                 .ok_or_else(|| {
-                    CompileError::codegen_error(format!("Cannot parse {} as int", value), location)
+                    Diagnostic::codegen_error(&format!("Cannot parse {} as int", value), location)
                 })
                 .map(BasicValueEnum::IntValue),
             BasicTypeEnum::FloatType { 0: float_type } => {
                 let value = float_type.const_float_from_string(value);
                 Ok(BasicValueEnum::FloatValue(value))
             }
-            _ => Err(CompileError::codegen_error(
+            _ => Err(Diagnostic::codegen_error(
                 "expected numeric type".into(),
                 location,
             )),
@@ -178,7 +178,7 @@ impl<'a> Llvm<'a> {
     }
 
     /// create a null pointer
-    pub fn create_null_ptr(&self) -> Result<BasicValueEnum<'a>, CompileError> {
+    pub fn create_null_ptr(&self) -> Result<BasicValueEnum<'a>, Diagnostic> {
         let itype = self.context.i32_type().ptr_type(AddressSpace::Generic);
         let value = itype.const_null();
         Ok(value.into())
@@ -190,7 +190,7 @@ impl<'a> Llvm<'a> {
     pub fn create_const_utf8_string(
         &self,
         value: &str,
-    ) -> Result<BasicValueEnum<'a>, CompileError> {
+    ) -> Result<BasicValueEnum<'a>, Diagnostic> {
         self.create_llvm_const_vec_string(value.as_bytes())
     }
 
@@ -200,7 +200,7 @@ impl<'a> Llvm<'a> {
     pub fn create_const_utf16_string(
         &self,
         value: &str,
-    ) -> Result<BasicValueEnum<'a>, CompileError> {
+    ) -> Result<BasicValueEnum<'a>, Diagnostic> {
         let mut utf16_chars: Vec<u16> = value.encode_utf16().collect();
         //it only contains a single NUL-terminator-byte so we add a second one
         utf16_chars.push(0);
@@ -213,7 +213,7 @@ impl<'a> Llvm<'a> {
     pub fn create_llvm_const_utf16_vec_string(
         &self,
         value: &[u16],
-    ) -> Result<BasicValueEnum<'a>, CompileError> {
+    ) -> Result<BasicValueEnum<'a>, Diagnostic> {
         let values: Vec<IntValue> = value
             .iter()
             .map(|it| self.context.i16_type().const_int(*it as u64, false))
@@ -227,7 +227,7 @@ impl<'a> Llvm<'a> {
     pub fn create_llvm_const_vec_string(
         &self,
         value: &[u8],
-    ) -> Result<BasicValueEnum<'a>, CompileError> {
+    ) -> Result<BasicValueEnum<'a>, Diagnostic> {
         let exp_value = self.context.const_string(value, true);
         Ok(BasicValueEnum::VectorValue(exp_value))
     }
