+++ src/codegen/generators/statement_generator.rs
@@ -2,18 +2,10 @@
 use super::{
     expression_generator::ExpressionCodeGenerator, llvm::Llvm, pou_generator::PouGenerator,
 };
-use crate::{
-    ast::{flatten_expression_list, AstStatement, ConditionalBlock, Operator, SourceRange},
-    codegen::llvm_typesystem,
-    codegen::LlvmTypedIndex,
-    compile_error::{CompileError, INTERNAL_LLVM_ERROR},
-    index::{ImplementationIndexEntry, Index},
-    resolver::AnnotationMap,
-    typesystem::{
+use crate::{ast::{flatten_expression_list, AstStatement, ConditionalBlock, Operator, SourceRange}, codegen::LlvmTypedIndex, codegen::llvm_typesystem, diagnostics::{Diagnostic, INTERNAL_LLVM_ERROR}, index::{ImplementationIndexEntry, Index}, resolver::AnnotationMap, typesystem::{
         DataTypeInformation, DINT_TYPE, RANGE_CHECK_LS_FN, RANGE_CHECK_LU_FN, RANGE_CHECK_S_FN,
         RANGE_CHECK_U_FN,
-    },
-};
+    }};
 use inkwell::{
     basic_block::BasicBlock,
     builder::Builder,
@@ -85,7 +77,7 @@ impl<'a, 'b> StatementCodeGenerator<'a, 'b> {
     }
 
     /// generates a list of statements
-    pub fn generate_body(&self, statements: &[AstStatement]) -> Result<(), CompileError> {
+    pub fn generate_body(&self, statements: &[AstStatement]) -> Result<(), Diagnostic> {
         for s in statements {
             self.generate_statement(s)?;
         }
@@ -108,7 +100,7 @@ impl<'a, 'b> StatementCodeGenerator<'a, 'b> {
     /// genertes a single statement
     ///
     /// - `statement` the statement to be generated
-    pub fn generate_statement(&self, statement: &AstStatement) -> Result<(), CompileError> {
+    pub fn generate_statement(&self, statement: &AstStatement) -> Result<(), Diagnostic> {
         match statement {
             AstStatement::EmptyStatement { .. } => {
                 //nothing to generate
@@ -159,10 +151,10 @@ impl<'a, 'b> StatementCodeGenerator<'a, 'b> {
                     self.llvm.builder.build_unconditional_branch(*exit_block);
                     self.generate_buffer_block();
                 } else {
-                    return Err(CompileError::CodeGenError {
-                        message: "Cannot break out of loop when not inside a loop".into(),
-                        location: location.clone(),
-                    });
+                    return Err(Diagnostic::codegen_error(
+                        "Cannot break out of loop when not inside a loop",
+                        location.clone(),
+                    ));
                 }
             }
             AstStatement::ContinueStatement { location, .. } => {
@@ -170,10 +162,10 @@ impl<'a, 'b> StatementCodeGenerator<'a, 'b> {
                     self.llvm.builder.build_unconditional_branch(*cont_block);
                     self.generate_buffer_block();
                 } else {
-                    return Err(CompileError::CodeGenError {
-                        message: "Cannot continue loop when not inside a loop".into(),
-                        location: location.clone(),
-                    });
+                    return Err(Diagnostic::codegen_error(
+                        "Cannot continue loop when not inside a loop",
+                        location.clone(),
+                    ));
                 }
             }
             _ => {
@@ -192,7 +184,7 @@ impl<'a, 'b> StatementCodeGenerator<'a, 'b> {
         &self,
         left_statement: &AstStatement,
         right_statement: &AstStatement,
-    ) -> Result<(), CompileError> {
+    ) -> Result<(), Diagnostic> {
         //TODO: Looks hacky, the strings will be similar so we should look into making the assignment a bit nicer.
         if left_statement.has_direct_access() {
             return self.generate_direct_access_assignment(left_statement, right_statement);
@@ -231,7 +223,7 @@ impl<'a, 'b> StatementCodeGenerator<'a, 'b> {
         &self,
         left_statement: &AstStatement,
         right_statement: &AstStatement,
-    ) -> Result<(), CompileError> {
+    ) -> Result<(), Diagnostic> {
         //TODO : Validation
         let exp_gen = self.create_expr_generator();
         if let AstStatement::QualifiedReference { elements, .. } = left_statement {
@@ -264,8 +256,8 @@ impl<'a, 'b> StatementCodeGenerator<'a, 'b> {
                         left_type,
                     )
                 } else {
-                    Err(CompileError::codegen_error(
-                        format!("{:?} not a direct access", element),
+                    Err(Diagnostic::syntax_error(
+                        &format!("{:?} not a direct access", element),
                         element.get_location(),
                     ))
                 }?;
@@ -278,8 +270,8 @@ impl<'a, 'b> StatementCodeGenerator<'a, 'b> {
                             left_type,
                         )
                     } else {
-                        Err(CompileError::codegen_error(
-                            format!("{:?} not a direct access", element),
+                        Err(Diagnostic::syntax_error(
+                            &format!("{:?} not a direct access", element),
                             element.get_location(),
                         ))
                     }?;
@@ -370,7 +362,7 @@ impl<'a, 'b> StatementCodeGenerator<'a, 'b> {
         end: &AstStatement,
         by_step: &Option<Box<AstStatement>>,
         body: &[AstStatement],
-    ) -> Result<(), CompileError> {
+    ) -> Result<(), Diagnostic> {
         let (builder, current_function, context) = self.get_llvm_deps();
         self.generate_assignment_statement(counter, start)?;
         let condition_check = context.append_basic_block(current_function, "condition_check");
@@ -454,7 +446,7 @@ impl<'a, 'b> StatementCodeGenerator<'a, 'b> {
         selector: &AstStatement,
         conditional_blocks: &[ConditionalBlock],
         else_body: &[AstStatement],
-    ) -> Result<Option<BasicValueEnum<'a>>, CompileError> {
+    ) -> Result<Option<BasicValueEnum<'a>>, Diagnostic> {
         let (builder, current_function, context) = self.get_llvm_deps();
         //Continue
         let continue_block = context.append_basic_block(current_function, "continue");
@@ -523,7 +515,7 @@ impl<'a, 'b> StatementCodeGenerator<'a, 'b> {
         start: &AstStatement,
         end: &AstStatement,
         match_block: BasicBlock,
-    ) -> Result<BasicBlock, CompileError> {
+    ) -> Result<BasicBlock, Diagnostic> {
         let (builder, _, context) = self.get_llvm_deps();
 
         let range_then = context.insert_basic_block_after(
@@ -566,7 +558,7 @@ impl<'a, 'b> StatementCodeGenerator<'a, 'b> {
         &self,
         condition: &AstStatement,
         body: &[AstStatement],
-    ) -> Result<Option<BasicValueEnum<'a>>, CompileError> {
+    ) -> Result<Option<BasicValueEnum<'a>>, Diagnostic> {
         let builder = &self.llvm.builder;
         let basic_block = builder.get_insert_block().expect(INTERNAL_LLVM_ERROR);
         self.generate_base_while_statement(condition, body)?;
@@ -596,7 +588,7 @@ impl<'a, 'b> StatementCodeGenerator<'a, 'b> {
         &self,
         condition: &AstStatement,
         body: &[AstStatement],
-    ) -> Result<Option<BasicValueEnum<'a>>, CompileError> {
+    ) -> Result<Option<BasicValueEnum<'a>>, Diagnostic> {
         let builder = &self.llvm.builder;
         let basic_block = builder.get_insert_block().expect(INTERNAL_LLVM_ERROR);
         self.generate_base_while_statement(condition, body)?;
@@ -618,7 +610,7 @@ impl<'a, 'b> StatementCodeGenerator<'a, 'b> {
         &self,
         condition: &AstStatement,
         body: &[AstStatement],
-    ) -> Result<Option<BasicValueEnum>, CompileError> {
+    ) -> Result<Option<BasicValueEnum>, Diagnostic> {
         let (builder, current_function, context) = self.get_llvm_deps();
         let condition_check = context.append_basic_block(current_function, "condition_check");
         let while_body = context.append_basic_block(current_function, "while_body");
@@ -661,7 +653,7 @@ impl<'a, 'b> StatementCodeGenerator<'a, 'b> {
         &self,
         conditional_blocks: &[ConditionalBlock],
         else_body: &[AstStatement],
-    ) -> Result<(), CompileError> {
+    ) -> Result<(), Diagnostic> {
         let (builder, current_function, context) = self.get_llvm_deps();
         let mut blocks = vec![builder.get_insert_block().expect(INTERNAL_LLVM_ERROR)];
         for _ in 1..conditional_blocks.len() {
