+++ src/lib.rs
@@ -18,15 +18,16 @@
 //! [`IEC61131-3`]: https://en.wikipedia.org/wiki/IEC_61131-3
 //! [`IR`]: https://llvm.org/docs/LangRef.html
 use std::fs;
-use std::ops::Range;
+
 use std::path::Path;
 
-use ast::{DataTypeDeclaration, PouType, SourceRange};
+use ast::{PouType, SourceRange};
 use codespan_reporting::diagnostic::{self, Label};
 use codespan_reporting::files::SimpleFiles;
 use codespan_reporting::term::termcolor::{ColorChoice, StandardStream};
 use codespan_reporting::term::{self, Chars, Styles};
 use compile_error::CompileError;
+use diagnostics::Diagnostic;
 use encoding_rs::Encoding;
 use encoding_rs_io::DecodeReaderBytesBuilder;
 use index::Index;
@@ -44,6 +45,7 @@ mod ast;
 pub mod cli;
 mod codegen;
 pub mod compile_error;
+mod diagnostics;
 pub mod index;
 mod lexer;
 mod parser;
@@ -55,281 +57,6 @@ mod validation;
 #[macro_use]
 #[cfg(test)]
 extern crate pretty_assertions;
-
-#[derive(PartialEq, Debug, Clone)]
-pub enum Diagnostic {
-    SyntaxError {
-        message: String,
-        range: SourceRange,
-        err_no: ErrNo,
-    },
-    ImprovementSuggestion {
-        message: String,
-        range: SourceRange,
-    },
-}
-
-#[allow(non_camel_case_types)]
-#[derive(PartialEq, Debug, Clone)]
-pub enum ErrNo {
-    undefined,
-
-    //syntax
-    syntax__generic_error,
-    syntax__missing_token,
-    syntax__unexpected_token,
-
-    //semantic
-    // pou related
-    pou__missing_return_type,
-    pou__unexpected_return_type,
-    pou__unsupported_return_type,
-    pou__empty_variable_block,
-
-    //variable related
-    var__unresolved_constant,
-    var__invalid_constant_block,
-    var__invalid_constant,
-    var__cannot_assign_to_const,
-
-    //reference related
-    reference__unresolved,
-
-    //type related
-    type__literal_out_of_range,
-    type__incompatible_literal_cast,
-    type__incompatible_directaccess,
-    type__incompatible_directaccess_variable,
-    type__incompatible_directaccess_range,
-    type__expected_literal,
-}
-
-impl Diagnostic {
-    pub fn syntax_error(message: &str, range: SourceRange) -> Diagnostic {
-        Diagnostic::SyntaxError {
-            message: message.to_string(),
-            range,
-            err_no: ErrNo::syntax__generic_error,
-        }
-    }
-
-    pub fn unexpected_token_found(expected: &str, found: &str, range: SourceRange) -> Diagnostic {
-        Diagnostic::SyntaxError {
-            message: format!(
-                "Unexpected token: expected {} but found {}",
-                expected, found
-            ),
-            range,
-            err_no: ErrNo::syntax__unexpected_token,
-        }
-    }
-
-    pub fn unexpected_initializer_on_function_return(range: SourceRange) -> Diagnostic {
-        Diagnostic::SyntaxError {
-            message: "Return types cannot have a default value".into(),
-            range,
-            err_no: ErrNo::syntax__unexpected_token,
-        }
-    }
-
-    pub fn return_type_not_supported(pou_type: &PouType, range: SourceRange) -> Diagnostic {
-        Diagnostic::SyntaxError {
-            message: format!(
-                "POU Type {:?} does not support a return type. Did you mean Function?",
-                pou_type
-            ),
-            range,
-            err_no: ErrNo::pou__unexpected_return_type,
-        }
-    }
-
-    pub fn function_unsupported_return_type(data_type: &DataTypeDeclaration) -> Diagnostic {
-        Diagnostic::SyntaxError {
-            message: format!(
-                "Data Type {:?} not supported as a function return type!",
-                data_type
-            ),
-            range: data_type.get_location(),
-            err_no: ErrNo::pou__unsupported_return_type,
-        }
-    }
-
-    pub fn function_return_missing(range: SourceRange) -> Diagnostic {
-        Diagnostic::SyntaxError {
-            message: "Function Return type missing".into(),
-            range,
-            err_no: ErrNo::pou__missing_return_type,
-        }
-    }
-
-    pub fn missing_token(epxected_token: &str, range: SourceRange) -> Diagnostic {
-        Diagnostic::SyntaxError {
-            message: format!("Missing expected Token {}", epxected_token),
-            range,
-            err_no: ErrNo::syntax__missing_token,
-        }
-    }
-
-    pub fn missing_action_container(range: SourceRange) -> Diagnostic {
-        Diagnostic::SyntaxError {
-            message: "Missing Actions Container Name".to_string(),
-            range,
-            err_no: ErrNo::undefined,
-        }
-    }
-
-    pub fn unrseolved_reference(reference: &str, location: SourceRange) -> Diagnostic {
-        Diagnostic::SyntaxError {
-            message: format!("Could not resolve reference to {:}", reference),
-            range: location,
-            err_no: ErrNo::reference__unresolved,
-        }
-    }
-
-    pub fn incompatible_directaccess(
-        access_type: &str,
-        access_size: u64,
-        location: SourceRange,
-    ) -> Diagnostic {
-        Diagnostic::SyntaxError {
-            message: format!(
-                "{}-Wise access requires a Numerical type larger than {} bits",
-                access_type, access_size
-            ),
-            range: location,
-            err_no: ErrNo::type__incompatible_directaccess,
-        }
-    }
-
-    pub fn incompatible_directaccess_range(
-        access_type: &str,
-        target_type: &str,
-        access_range: Range<u64>,
-        location: SourceRange,
-    ) -> Diagnostic {
-        Diagnostic::SyntaxError {
-            message: format!(
-                "{}-Wise access for type {} must be in the range {}..{}",
-                access_type, target_type, access_range.start, access_range.end
-            ),
-            range: location,
-            err_no: ErrNo::type__incompatible_directaccess_range,
-        }
-    }
-
-    pub fn incompatible_directaccess_variable(
-        access_type: &str,
-        location: SourceRange,
-    ) -> Diagnostic {
-        Diagnostic::SyntaxError {
-            message: format!(
-                "Invalid type {} for direct variable access. Only variables of Integer types are allowed",
-                access_type
-            ),
-            range: location,
-            err_no: ErrNo::type__incompatible_directaccess_variable,
-        }
-    }
-
-    pub fn incompatible_literal_cast(
-        cast_type: &str,
-        literal_type: &str,
-        location: SourceRange,
-    ) -> Diagnostic {
-        Diagnostic::SyntaxError {
-            message: format!(
-                "Literal {:} is not campatible to {:}",
-                literal_type, cast_type
-            ),
-            range: location,
-            err_no: ErrNo::type__incompatible_literal_cast,
-        }
-    }
-
-    pub fn literal_expected(location: SourceRange) -> Diagnostic {
-        Diagnostic::SyntaxError {
-            message: "Expected literal".into(),
-            range: location,
-            err_no: ErrNo::type__expected_literal,
-        }
-    }
-
-    pub fn literal_out_of_range(
-        literal: &str,
-        range_hint: &str,
-        location: SourceRange,
-    ) -> Diagnostic {
-        Diagnostic::SyntaxError {
-            message: format!("Literal {:} out of range ({})", literal, range_hint),
-            range: location,
-            err_no: ErrNo::type__literal_out_of_range,
-        }
-    }
-
-    pub fn empty_variable_block(location: SourceRange) -> Diagnostic {
-        Diagnostic::SyntaxError {
-            message: "Variable block is empty".into(),
-            range: location,
-            err_no: ErrNo::pou__empty_variable_block,
-        }
-    }
-
-    pub fn unresolved_constant(
-        constant_name: &str,
-        reason: Option<&str>,
-        location: SourceRange,
-    ) -> Diagnostic {
-        Diagnostic::SyntaxError {
-            message: format!(
-                "Unresolved constant '{:}' variable{:}",
-                constant_name,
-                reason
-                    .map(|it| format!(": {:}", it))
-                    .unwrap_or_else(|| "".into()),
-            ),
-            range: location,
-            err_no: ErrNo::pou__empty_variable_block,
-        }
-    }
-
-    pub fn invalid_constant_block(location: SourceRange) -> Diagnostic {
-        Diagnostic::SyntaxError {
-            message: "This variable block does not support the CONSTANT modifier".to_string(),
-            range: location,
-            err_no: ErrNo::var__invalid_constant_block,
-        }
-    }
-
-    pub fn invalid_constant(constant_name: &str, location: SourceRange) -> Diagnostic {
-        Diagnostic::SyntaxError {
-            message: format!("Invalid constant {:} - Functionblock- and Class-instances cannot be delcared constant", constant_name),
-            range: location,
-            err_no: ErrNo::var__invalid_constant,
-        }
-    }
-
-    pub fn cannot_assign_to_constant(qualified_name: &str, location: SourceRange) -> Diagnostic {
-        Diagnostic::SyntaxError {
-            message: format!("Cannot assign to CONSTANT '{:}'", qualified_name),
-            range: location,
-            err_no: ErrNo::var__cannot_assign_to_const,
-        }
-    }
-
-    pub fn get_message(&self) -> &str {
-        match self {
-            Diagnostic::SyntaxError { message, .. } => message.as_str(),
-            Diagnostic::ImprovementSuggestion { message, .. } => message.as_str(),
-        }
-    }
-
-    pub fn get_location(&self) -> SourceRange {
-        match self {
-            Diagnostic::SyntaxError { range, .. } => range.clone(),
-            Diagnostic::ImprovementSuggestion { range, .. } => range.clone(),
-        }
-    }
-}
 
 /// SourceContainers offer source-code to be compiled via the load_source function.
 /// Furthermore it offers a location-String used when reporting diagnostics.
